Pull out rest of record into specific files
